refactor(js): extract loader functions to `components/loader.js` module
Steps:
- Create `assets/js/components/loader.js` with `loaderAdd` and `loaderRemove` exports
- Keep `loaderCreate` as internal function (not exported)
- Remove loader functions from `scripts.js`
- Add `import { loaderAdd, loaderRemove }` to `scripts.js`

Reason:
- Part of the modularization effort to split `scripts.js` into separate ES modules

References:
- Issue #104
